Event Planning Assistant/requirements.txt
streamlit
langchain
langchain-openai
langchain_openai
langchain_community
geocoder
httpx
python-dotenv
requests
sqlite3-binary
